test: moved mountWithCartProvider to its own file to keep fixtures cleaner
