+++ packages/hydrogen/src/components/CartLineQuantityAdjustButton/tests/CartLineQuantityAdjustButton.test.tsx
@@ -4,10 +4,8 @@ import {CartLineQuantity} from '../../CartLineQuantity';
 import {CartLineQuantityAdjustButton} from '../CartLineQuantityAdjustButton';
 import {CART_LINE} from '../../CartLineProvider/tests/fixtures';
 import {useCart} from '../../CartProvider';
-import {
-  CART_WITH_LINES_FLATTENED,
-  mountWithCartProvider,
-} from '../../CartProvider/tests/fixtures';
+import {CART_WITH_LINES_FLATTENED} from '../../CartProvider/tests/fixtures';
+import {mountWithCartProvider} from '../../CartProvider/tests/CartProvider';
 
 describe('CartLineQuantityAdjustButton', () => {
   it('increases quantity', () => {
